# src/Api/Controllers/CategoriesController.cs
using Kart.Shared.Observability;
using KartCategoryService.Api.Common;
using KartCategoryService.Api.Security;
using KartCategoryService.Application.Common.Models;
using KartCategoryService.Application.Features.CreateCategory;
using KartCategoryService.Application.Features.DeprecateCategory;
using KartCategoryService.Application.Features.GetCategory;
using KartCategoryService.Application.Features.ListCategories;
using KartCategoryService.Application.Features.MoveCategory;
using KartCategoryService.Application.Features.RenameCategory;
using KartCategoryService.Application.Features.ReorderCategory;
using MediatR;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;

namespace KartCategoryService.Api.Controllers;

/// <summary>Every write action here belongs to the "Category & Attribute Management (Admin)" flow - KartFlowContext.Push wraps each one so every downstream log line (handler, persistence, outbox) inherits the Flow tag. GET/list is unauthenticated public catalog browsing, not an admin-management action, so it's deliberately left untagged.</summary>
[ApiController]
[Route("v1/categories")]
public sealed class CategoriesController : ControllerBase
{
    private const string FlowName = "CategoryAttributeManagementAdmin";

    private readonly ISender _sender;
    private readonly ILogger<CategoriesController> _logger;

    public CategoriesController(ISender sender, ILogger<CategoriesController> logger)
    {
        _sender = sender;
        _logger = logger;
    }

    /// <summary>api-contract.yaml listCategories - GET /v1/categories. CanRead is unconditional (ddd-model.md).</summary>
    [HttpGet]
    [ProducesResponseType(typeof(IReadOnlyList<CategoryDto>), StatusCodes.Status200OK)]
    public async Task<ActionResult<IReadOnlyList<CategoryDto>>> ListCategories(
        [FromQuery] Guid? parentId,
        [FromQuery] bool includeDeprecated,
        CancellationToken cancellationToken)
    {
        var categories = await _sender.Send(new ListCategoriesQuery(parentId, includeDeprecated), cancellationToken);
        return Ok(categories);
    }

    /// <summary>api-contract.yaml getCategory - GET /v1/categories/{categoryId}. Single-category
    /// lookup by id (e.g. a storefront category page's own title) - CanRead is unconditional
    /// (ddd-model.md), same as ListCategories. 404 if not found or deprecated.</summary>
    [HttpGet("{categoryId:guid}")]
    [ProducesResponseType(typeof(CategoryDto), StatusCodes.Status200OK)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status404NotFound)]
    public async Task<ActionResult<CategoryDto>> GetCategory([FromRoute] Guid categoryId, CancellationToken cancellationToken)
    {
        var category = await _sender.Send(new GetCategoryQuery(categoryId), cancellationToken);
        return category is null ? NotFound() : Ok(category);
    }

    /// <summary>api-contract.yaml createCategory - POST /v1/categories (RBAC-gated, Admin only).</summary>
    [HttpPost]
    [Authorize(Policy = AuthenticationExtensions.AdminPolicy)]
    [ProducesResponseType(typeof(CategoryDto), StatusCodes.Status201Created)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status404NotFound)]
    public async Task<ActionResult<CategoryDto>> CreateCategory(
        [FromBody] CreateCategoryRequest request,
        CancellationToken cancellationToken)
    {
        using var _ = KartFlowContext.Push(FlowName);
        _logger.LogInformation("Stage {Stage}: create-category request received (parentId {ParentId})", "CategoryCreateRequestReceived", request.ParentId);

        var command = new CreateCategoryCommand(request.Name, request.ParentId);
        var result = await _sender.Send(command, cancellationToken);
        return this.ToActionResult<CategoryDto, CategoryDto>(
            result,
            category => CreatedAtAction(nameof(ListCategories), new { parentId = category.ParentId }, category));
    }

    /// <summary>api-contract.yaml renameCategory - PATCH /v1/categories/{categoryId} (RBAC-gated, Admin only).</summary>
    [HttpPatch("{categoryId:guid}")]
    [Authorize(Policy = AuthenticationExtensions.AdminPolicy)]
    [ProducesResponseType(typeof(CategoryDto), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status404NotFound)]
    public async Task<ActionResult<CategoryDto>> RenameCategory(
        [FromRoute] Guid categoryId,
        [FromBody] RenameCategoryRequest request,
        CancellationToken cancellationToken)
    {
        using var _ = KartFlowContext.Push(FlowName);
        _logger.LogInformation("Stage {Stage}: rename-category request received (categoryId {CategoryId})", "CategoryRenameRequestReceived", categoryId);

        var command = new RenameCategoryCommand(categoryId, request.Name);
        var result = await _sender.Send(command, cancellationToken);
        return this.ToActionResult<CategoryDto, CategoryDto>(result, category => Ok(category));
    }

    /// <summary>api-contract.yaml reorderCategory - POST /v1/categories/{categoryId}/reorder (RBAC-gated, Admin only). Previously missing entirely - kart-admin-service's ReorderCategoryCommand has always called this exact route, so it 404'd on every reorder attempt until this endpoint was added.</summary>
    [HttpPost("{categoryId:guid}/reorder")]
    [Authorize(Policy = AuthenticationExtensions.AdminPolicy)]
    [ProducesResponseType(typeof(CategoryDto), StatusCodes.Status200OK)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status404NotFound)]
    public async Task<ActionResult<CategoryDto>> ReorderCategory(
        [FromRoute] Guid categoryId,
        [FromBody] ReorderCategoryRequest request,
        CancellationToken cancellationToken)
    {
        using var _ = KartFlowContext.Push(FlowName);
        _logger.LogInformation("Stage {Stage}: reorder-category request received (categoryId {CategoryId}, displayOrder {DisplayOrder})", "CategoryReorderRequestReceived", categoryId, request.DisplayOrder);

        var command = new ReorderCategoryCommand(categoryId, request.DisplayOrder);
        var result = await _sender.Send(command, cancellationToken);
        return this.ToActionResult<CategoryDto, CategoryDto>(result, category => Ok(category));
    }

    /// <summary>api-contract.yaml moveCategory - POST /v1/categories/{categoryId}/move (RBAC-gated, Admin only).</summary>
    [HttpPost("{categoryId:guid}/move")]
    [Authorize(Policy = AuthenticationExtensions.AdminPolicy)]
    [ProducesResponseType(typeof(CategoryDto), StatusCodes.Status200OK)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status404NotFound)]
    public async Task<ActionResult<CategoryDto>> MoveCategory(
        [FromRoute] Guid categoryId,
        [FromBody] MoveCategoryRequest request,
        CancellationToken cancellationToken)
    {
        using var _ = KartFlowContext.Push(FlowName);
        _logger.LogInformation("Stage {Stage}: move-category request received (categoryId {CategoryId})", "CategoryMoveRequestReceived", categoryId);

        var command = new MoveCategoryCommand(categoryId, request.NewParentId);
        var result = await _sender.Send(command, cancellationToken);
        return this.ToActionResult<CategoryDto, CategoryDto>(result, category => Ok(category));
    }

    /// <summary>api-contract.yaml deprecateCategory - DELETE /v1/categories/{categoryId} (RBAC-gated, Admin only).</summary>
    [HttpDelete("{categoryId:guid}")]
    [Authorize(Policy = AuthenticationExtensions.AdminPolicy)]
    [ProducesResponseType(StatusCodes.Status204NoContent)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    [ProducesResponseType(typeof(ProblemDto), StatusCodes.Status404NotFound)]
    public async Task<IActionResult> DeprecateCategory([FromRoute] Guid categoryId, CancellationToken cancellationToken)
    {
        using var _ = KartFlowContext.Push(FlowName);
        _logger.LogInformation("Stage {Stage}: deprecate-category request received (categoryId {CategoryId})", "CategoryDeprecateRequestReceived", categoryId);

        var command = new DeprecateCategoryCommand(categoryId);
        var result = await _sender.Send(command, cancellationToken);
        return result.IsSuccess ? NoContent() : this.MapFailure(result.Error);
    }
}

/// <summary>api-contract.yaml createCategory requestBody shape.</summary>
public sealed record CreateCategoryRequest(string Name, Guid? ParentId);

/// <summary>api-contract.yaml renameCategory requestBody shape.</summary>
public sealed record RenameCategoryRequest(string Name);

/// <summary>api-contract.yaml reorderCategory requestBody shape.</summary>
public sealed record ReorderCategoryRequest(int DisplayOrder);

/// <summary>api-contract.yaml moveCategory requestBody shape.</summary>
public sealed record MoveCategoryRequest(Guid? NewParentId);